refactor: move the menu HTML into 1 reusable file
