+++ pages/pray/action-global-map.php
@@ -111,6 +111,8 @@ class PG_Global_Prayer_App_Map extends PG_Global_Prayer_App {
         <script src="<?php echo esc_url( trailingslashit( plugin_dir_url( __FILE__ ) ) ) ?>report.js?ver=<?php echo esc_attr( fileatime( trailingslashit( plugin_dir_path( __FILE__ ) ) . 'report.js' ) ) ?>"></script>
         <script src="<?php echo esc_url( trailingslashit( plugin_dir_url( __DIR__ ) ) ) ?>assets/js/share.js?ver=<?php echo esc_attr( fileatime( trailingslashit( plugin_dir_path( __DIR__ ) ) ) . 'assets/js/share.js' ) ?>"></script>
         <?php
+
+        pg_toggle_user_elements();
     }
 
     public function footer_javascript() {
